requirements.txt
asgiref==3.6.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
boto3==1.26.115
botocore==1.29.115
dj-database-url==1.3.0
Django==4.1.7
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.4.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
whitenoise==6.4.0
